try to refactor samd11 code so samd21 can reuse
