+++ platform/samd11/sys_init.c
@@ -1,0 +1,35 @@
+// SPDX-License-Identifier: BSD-3-Clause
+// Copyright (c) 2017-2022, Alex Taradov <alex@taradov.com>. All rights reserved.
+
+//-----------------------------------------------------------------------------
+#include "samd11.h"
+#include "nvm_data.h"
+
+void sys_init(void)
+{
+  uint32_t coarse, fine;
+
+  NVMCTRL->CTRLB.reg = NVMCTRL_CTRLB_RWS(1);
+
+  SYSCTRL->INTFLAG.reg = SYSCTRL_INTFLAG_BOD33RDY | SYSCTRL_INTFLAG_BOD33DET |
+      SYSCTRL_INTFLAG_DFLLRDY;
+
+  coarse = NVM_READ_CAL(NVM_DFLL48M_COARSE_CAL);
+  fine = NVM_READ_CAL(NVM_DFLL48M_FINE_CAL);
+
+  SYSCTRL->DFLLCTRL.reg = 0; // See Errata 9905
+  while (0 == (SYSCTRL->PCLKSR.reg & SYSCTRL_PCLKSR_DFLLRDY));
+
+  SYSCTRL->DFLLMUL.reg = SYSCTRL_DFLLMUL_MUL(48000);
+  SYSCTRL->DFLLVAL.reg = SYSCTRL_DFLLVAL_COARSE(coarse) | SYSCTRL_DFLLVAL_FINE(fine);
+
+  SYSCTRL->DFLLCTRL.reg = SYSCTRL_DFLLCTRL_ENABLE | SYSCTRL_DFLLCTRL_USBCRM |
+      SYSCTRL_DFLLCTRL_MODE | SYSCTRL_DFLLCTRL_CCDIS;
+
+  while (0 == (SYSCTRL->PCLKSR.reg & SYSCTRL_PCLKSR_DFLLRDY));
+
+  GCLK->GENCTRL.reg = GCLK_GENCTRL_ID(0) | GCLK_GENCTRL_SRC(GCLK_SOURCE_DFLL48M) |
+      GCLK_GENCTRL_RUNSTDBY | GCLK_GENCTRL_GENEN;
+  while (GCLK->STATUS.reg & GCLK_STATUS_SYNCBUSY);
+}
+
