+++ platform/samdx1/dap_config.h
@@ -1,38 +1,27 @@
 // SPDX-License-Identifier: BSD-3-Clause
-// Copyright (c) 2016-2022, Alex Taradov <alex@taradov.com>. All rights reserved.
+// Copyright (c) 2017-2022, Alex Taradov <alex@taradov.com>. All rights reserved.
 
 #ifndef _DAP_CONFIG_H_
 #define _DAP_CONFIG_H_
 
 /*- Includes ----------------------------------------------------------------*/
-#include "samd21.h"
-#include "hal_gpio.h"
+#include "samdx1.h"
+#include "hal_config.h"
 
 /*- Definitions -------------------------------------------------------------*/
-HAL_GPIO_PIN(SWCLK_TCK,    B, 11)
-HAL_GPIO_PIN(SWDIO_TMS,    B, 17)
-HAL_GPIO_PIN(TDI,          A, 8)
-HAL_GPIO_PIN(TDO,          A, 9)
-HAL_GPIO_PIN(nTRST,        B, 23)
-HAL_GPIO_PIN(nRESET,       A, 2)
-
-HAL_GPIO_PIN(LED,          B, 30)
-
-#define DAP_CONFIG_ENABLE_JTAG
-
 #define DAP_CONFIG_DEFAULT_PORT        DAP_PORT_SWD
 #define DAP_CONFIG_DEFAULT_CLOCK       1000000 // Hz
 
 #define DAP_CONFIG_PACKET_SIZE         64
-#define DAP_CONFIG_PACKET_COUNT        1
+#define DAP_CONFIG_PACKET_COUNT        2
 
 #define DAP_CONFIG_JTAG_DEV_COUNT      8
 
 // DAP_CONFIG_PRODUCT_STR must contain "CMSIS-DAP" to be compatible with the standard
 #define DAP_CONFIG_VENDOR_STR          "Alex Taradov"
 #define DAP_CONFIG_PRODUCT_STR         "Generic CMSIS-DAP Adapter"
-#define DAP_CONFIG_SER_NUM_STR         "123456"
-#define DAP_CONFIG_CMSIS_DAP_VER_STR   "1.3.0"
+#define DAP_CONFIG_SER_NUM_STR         usb_serial_number
+#define DAP_CONFIG_CMSIS_DAP_VER_STR   "2.0.0"
 
 //#define DAP_CONFIG_RESET_TARGET_FN     target_specific_reset_function
 //#define DAP_CONFIG_VENDOR_FN           vendor_command_handler_function
@@ -41,11 +30,14 @@ HAL_GPIO_PIN(LED,          B, 30)
 #define DAP_CONFIG_PERFORMANCE_ATTR    __attribute__((section(".ramfunc")))
 
 // A value at which dap_clock_test() produces 1 kHz output on the SWCLK pin
-#define DAP_CONFIG_DELAY_CONSTANT      4700
+#define DAP_CONFIG_DELAY_CONSTANT      7700
 
 // A threshold for switching to fast clock (no added delays)
 // This is the frequency produced by dap_clock_test(1) on the SWCLK pin
-#define DAP_CONFIG_FAST_CLOCK          3600000 // Hz
+#define DAP_CONFIG_FAST_CLOCK          2400000 // Hz
+
+/*- Prototypes --------------------------------------------------------------*/
+extern char usb_serial_number[16];
 
 /*- Implementations ---------------------------------------------------------*/
 
@@ -64,13 +56,21 @@ static inline void DAP_CONFIG_SWDIO_TMS_write(int value)
 //-----------------------------------------------------------------------------
 static inline void DAP_CONFIG_TDI_write(int value)
 {
+#ifdef DAP_CONFIG_ENABLE_JTAG
   HAL_GPIO_TDI_write(value);
+#else
+  (void)value;
+#endif
 }
 
 //-----------------------------------------------------------------------------
 static inline void DAP_CONFIG_TDO_write(int value)
 {
+#ifdef DAP_CONFIG_ENABLE_JTAG
   HAL_GPIO_TDO_write(value);
+#else
+  (void)value;
+#endif
 }
 
 //-----------------------------------------------------------------------------
@@ -100,13 +100,21 @@ static inline int DAP_CONFIG_SWDIO_TMS_read(void)
 //-----------------------------------------------------------------------------
 static inline int DAP_CONFIG_TDO_read(void)
 {
+#ifdef DAP_CONFIG_ENABLE_JTAG
   return HAL_GPIO_TDO_read();
+#else
+  return 0;
+#endif
 }
 
 //-----------------------------------------------------------------------------
 static inline int DAP_CONFIG_TDI_read(void)
 {
+#ifdef DAP_CONFIG_ENABLE_JTAG
   return HAL_GPIO_TDI_read();
+#else
+  return 0;
+#endif
 }
 
 //-----------------------------------------------------------------------------
@@ -151,8 +159,10 @@ static inline void DAP_CONFIG_SETUP(void)
   HAL_GPIO_SWCLK_TCK_in();
   HAL_GPIO_SWDIO_TMS_in();
   HAL_GPIO_nRESET_in();
+#ifdef DAP_CONFIG_ENABLE_JTAG
   HAL_GPIO_TDO_in();
   HAL_GPIO_TDI_in();
+#endif
 }
 
 //-----------------------------------------------------------------------------
@@ -161,8 +171,10 @@ static inline void DAP_CONFIG_DISCONNECT(void)
   HAL_GPIO_SWCLK_TCK_in();
   HAL_GPIO_SWDIO_TMS_in();
   HAL_GPIO_nRESET_in();
+#ifdef DAP_CONFIG_ENABLE_JTAG
   HAL_GPIO_TDO_in();
   HAL_GPIO_TDI_in();
+#endif
 }
 
 //-----------------------------------------------------------------------------
@@ -177,8 +189,10 @@ static inline void DAP_CONFIG_CONNECT_SWD(void)
   HAL_GPIO_nRESET_out();
   HAL_GPIO_nRESET_set();
 
+#ifdef DAP_CONFIG_ENABLE_JTAG
   HAL_GPIO_TDO_in();
   HAL_GPIO_TDI_in();
+#endif
 }
 
 //-----------------------------------------------------------------------------
@@ -190,20 +204,22 @@ static inline void DAP_CONFIG_CONNECT_JTAG(void)
   HAL_GPIO_SWCLK_TCK_out();
   HAL_GPIO_SWCLK_TCK_set();
 
+  HAL_GPIO_nRESET_out();
+  HAL_GPIO_nRESET_set();
+
+#ifdef DAP_CONFIG_ENABLE_JTAG
   HAL_GPIO_TDO_in();
 
   HAL_GPIO_TDI_out();
   HAL_GPIO_TDI_set();
-
-  HAL_GPIO_nRESET_out();
-  HAL_GPIO_nRESET_set();
+#endif
 }
 
 //-----------------------------------------------------------------------------
 static inline void DAP_CONFIG_LED(int index, int state)
 {
-  if (0 == index)
-    HAL_GPIO_LED_write(!state);
+  (void)index;
+  (void)state;
 }
 
 //-----------------------------------------------------------------------------
@@ -218,4 +234,3 @@ static inline void DAP_CONFIG_DELAY(uint32_t cycles)
 }
 
 #endif // _DAP_CONFIG_H_
-
