+++ platform/samdx1/main.c
@@ -7,7 +7,7 @@
 #include <stdbool.h>
 #include <stdalign.h>
 #include <string.h>
-#include "samd11.h"
+#include "samdx1.h"
 #include "hal_config.h"
 #include "nvm_data.h"
 #include "usb.h"
@@ -48,34 +48,7 @@ static bool app_vcp_open = false;
 
 /*- Implementations ---------------------------------------------------------*/
 
-//-----------------------------------------------------------------------------
-static void sys_init(void)
-{
-  uint32_t coarse, fine;
-
-  NVMCTRL->CTRLB.reg = NVMCTRL_CTRLB_RWS(1);
-
-  SYSCTRL->INTFLAG.reg = SYSCTRL_INTFLAG_BOD33RDY | SYSCTRL_INTFLAG_BOD33DET |
-      SYSCTRL_INTFLAG_DFLLRDY;
-
-  coarse = NVM_READ_CAL(NVM_DFLL48M_COARSE_CAL);
-  fine = NVM_READ_CAL(NVM_DFLL48M_FINE_CAL);
-
-  SYSCTRL->DFLLCTRL.reg = 0; // See Errata 9905
-  while (0 == (SYSCTRL->PCLKSR.reg & SYSCTRL_PCLKSR_DFLLRDY));
-
-  SYSCTRL->DFLLMUL.reg = SYSCTRL_DFLLMUL_MUL(48000);
-  SYSCTRL->DFLLVAL.reg = SYSCTRL_DFLLVAL_COARSE(coarse) | SYSCTRL_DFLLVAL_FINE(fine);
-
-  SYSCTRL->DFLLCTRL.reg = SYSCTRL_DFLLCTRL_ENABLE | SYSCTRL_DFLLCTRL_USBCRM |
-      SYSCTRL_DFLLCTRL_MODE | SYSCTRL_DFLLCTRL_CCDIS;
-
-  while (0 == (SYSCTRL->PCLKSR.reg & SYSCTRL_PCLKSR_DFLLRDY));
-
-  GCLK->GENCTRL.reg = GCLK_GENCTRL_ID(0) | GCLK_GENCTRL_SRC(GCLK_SOURCE_DFLL48M) |
-      GCLK_GENCTRL_RUNSTDBY | GCLK_GENCTRL_GENEN;
-  while (GCLK->STATUS.reg & GCLK_STATUS_SYNCBUSY);
-}
+extern void sys_init(void);
 
 //-----------------------------------------------------------------------------
 static void serial_number_init(void)
