+++ platform/samdx1/uart.h
