+++ platform/samdx1/usb_descriptors.c
